+++ CHANGELOG.md
@@ -18,6 +18,12 @@
 
 - Improved responsiveness of webinterface
 
+### API
+
+- Renamed `revocation_id` to `mom_id`
+- Restructured capabilities related to other mytokens
+- Added possibility to obtain history information for children and other tokens (capability)
+
 ## mytoken 0.6.1
 
 ### API